Fix timeout when using XCom with KubernetesPodOperator (#15388)
* Fix timeout when using XCom with KubernetesPodOperator

Currently, the xcom sidecar container for the KubernetesPodOperator will
sleep for 30 seconds before checking if the xcom has completed. This is
far too long of a wait, as a 1 second wait will ensure that the process
is not consistently blocked.

* get rid of unused variable

* remove unecessary poddefaults

* update docs

* Add UPDATING

* more description

* fix tests

* fix tests

* Finallyfound it
